Restructured files
Split Tileset impl across multiple files since the single file was
growing quite large

This also allows us to better separate by feature
